HUGE update, wieuw
Music will be used later on, hint hint
