--- auth.json.example
@@ -1,10 +1,0 @@
-{
-	"email" : "email of the discord account here",
-	"password" : "password of the discord account here",
-	"youtube_api_key": "create one here https://console.developers.google.com",
-	"imgflip_username": "https://imgflip.com/ username",
-	"imgflip_password": "https://imgflip.com/ password",
-	"wolfram_api_key": "go here and click sign up http://products.wolframalpha.com/api/",
-	"myapifilms_token": "http://api.myapifilms.com/requestToken.do",
-	"steam_api_key": "http://steamcommunity.com/dev/apikey"
-}
