+++ config.json.example
@@ -1,0 +1,15 @@
+{
+    "discord_email" : "E-mail adres from the bots Discord account",
+    "discord_password" : "Password from the bots Discord account",
+    "youtube_api_key": "Create one here https://console.developers.google.com",
+    "imgflip_username": "Imgflip username",
+    "imgflip_password": "Imgflip password",
+    "wolfram_api_key": "Optional, not currently needed",
+		"myapifilms_token": "http://api.myapifilms.com/requestToken.do",
+		"steam_api_key": "http://steamcommunity.com/dev/apikey",
+    "command_prefix": "Enter a symbol you want to use as the command prefix",
+    "join_servers_on_startup": "Enter a instant-invite here that you want the bot to join on start-up",
+    "admin_ids": ["108125505714139136", "107904023901777920", "110147170740494336"],
+    "image_folder": "./images",
+    "music_folder": "./music"
+}
