+++ runtime/abbreviations.json
@@ -1,0 +1,31 @@
+{
+	"cs": "Counter-Strike",
+	"hon": "Heroes of Newerth",
+	"hots": "Heroes of the Storm",
+	"sc2": "Starcraft II",
+	"wf": "Warframe",
+	"gtao": "Grand Theft Auto: Online",
+	"gta": "Grand Theft Auto",
+	"lol": "League of Legends",
+	"wow": "World of Warcraft",
+	"tf2": "Team Fortress 2",
+	"p2": "Portal 2",
+	"civ": "Civilization",
+	"se": "Space Engineers",
+	"cod": "Call of Duty",
+    "db": "Dirty Bomb",
+    "rs": "RuneScape",
+    "sr": "Shadowrun",
+    "mgs5": "Metal Gear Solid V",
+    "ed": "Elite: Dangerous",
+    "pd": "PayDay",
+    "pd2": "PayDay 2",
+    "me": "Medieval Engineers",
+    "me3": "Mass Effect 3",
+    "ws": "WildStar",
+    "aoe": "Age Of Empires",
+    "wt": "War Thunder",
+    "jc": "Just Cause",
+    "wd": "Watch_Dogs",
+    "sb": "StarBound"
+}
