+++ runtime/memes.json
@@ -1,0 +1,20 @@
+{
+ "brace": 61546,
+ "mostinteresting": 61532,
+ "fry": 61520,
+ "onedoesnot": 61579,
+ "yuno": 61527,
+ "success": 61544,
+ "allthethings": 61533,
+ "doge": 8072285,
+ "drevil": 40945639,
+ "skeptical": 101711,
+ "notime": 442575,
+ "yodawg": 101716,
+ "ermahgerd": 101462,
+ "hipsterariel": 86601,
+ "imagination": 163573,
+ "grumpycat": 405658,
+ "morpheus": 100947,
+ "1stworldproblems": 61539
+}
